aws_cdk_custom_vpc/requirements.txt
-e .
aws-cdk.core
aws_cdk.aws_ec2
aws_cdk.aws_s3